requirements.txt
asgiref==3.3.2
Django>=3.2.0,<3.3
django-debug-toolbar>=3.2
Pillow>=5.3.0,<5.4.0
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary==2.8.6
gunicorn>=20.1.0
django-autoslug>=1.9.8
pytest-django>=4.1.0
coverage>=5.5
pycodestyle
pytest-cov>=2.11.1
pylint>=2.7.4
pylint-django>=2.4.3
pylint-pytest>=1.1.1
pre-commit>=2.12.0
djangorestframework>=3.12.4,<3.13
dj-rest-auth
django-allauth
